+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+* Renamed `Disqualified::Record#unqueue` to `Disqualified::Record#unclaim`
+
+### Fixed
+
+* `Disqualified::Record#run!` now tears down the record for future retries
+
 ## v0.4.0 - 2024-08-10
 
 ### Added